Refactor out premature rounding of inter-stop-distances to allow for a more transparent code flow and debugging experience.

+++ src/main/kotlin/fi/hsl/jore4/hastus/data/jore/JoreDistanceBetweenTwoStopPoints.kt
@@ -10,5 +10,5 @@ package fi.hsl.jore4.hastus.data.jore
 data class JoreDistanceBetweenTwoStopPoints(
     val startLabel: String,
     val endLabel: String,
-    val distance: Int
+    val distance: Double
 )
